+++ src/main/java/org/dwcj/controls/AbstractDwcControl.java
@@ -27,7 +27,9 @@ public abstract class AbstractDwcControl implements IControl {
     private IThemable.Theme theme;
     private IExpansible.Expanse expanse;
     private Boolean visible = null;
+    private Boolean enabled = null;
     private String tooltipText = "";
+    
 
     protected static final BasisNumber BASISNUMBER_1 = BasisNumber.createBasisNumber(1);
     protected static final BasisNumber BASISNUMBER_25 = BasisNumber.createBasisNumber(25);
@@ -240,26 +242,26 @@ public abstract class AbstractDwcControl implements IControl {
         return tooltipText;
     }
 
-    // @Override
-    // public IControl setEnabled(boolean b) {
-    //     if (this.ctrl != null) try {
-    //         ctrl.setEnabled(b);
-    //     } catch (BBjException e) {
-    //         e.printStackTrace();
-    //     }
-    //     this.enabled = b;
-    //     return this;
-    // }
+    @Override
+    public IControl setEnabled(boolean b) {
+        if (this.ctrl != null) try {
+            ctrl.setEnabled(b);
+        } catch (BBjException e) {
+            e.printStackTrace();
+        }
+        this.enabled = b;
+        return this;
+    }
 
-    // @Override
-    // public boolean isEnabled() {
-    //     if (this.ctrl != null) try {
-    //         return ctrl.isEnabled();
-    //     } catch (BBjException e) {
-    //         e.printStackTrace();
-    //     }
-    //     return enabled;
-    // }
+    @Override
+    public boolean isEnabled() {
+        if (this.ctrl != null) try {
+            return ctrl.isEnabled();
+        } catch (BBjException e) {
+            e.printStackTrace();
+        }
+        return enabled;
+    }
 
     @Override
     public IControl setVisible(boolean b) {
