--- src/main/java/org/dwcj/controls/IEnable.java
@@ -1,9 +1,0 @@
-package org.dwcj.controls;
-
-public interface IEnable {
-
-    IEnable setEnabled(boolean enabled);
-
-    boolean isEnabled();
-
-}
