+++ src/main/java/org/dwcj/controls/TextComboBox.java
@@ -13,7 +13,7 @@ import java.util.Map;
 /**
  * ComboBoxEdit Control
  */
-public final class TextComboBox extends AbstractDwclistControl implements IThemable, IExpansible, IEnable {
+public final class TextComboBox extends AbstractDwclistControl implements IThemable, IExpansible {
 
     private BBjListEdit bbjListEdit;
 
@@ -207,23 +207,23 @@ public final class TextComboBox extends AbstractDwclistControl implements IThema
         return this;
     }
 
-    @Override
-    public TextComboBox setEnabled(boolean enabled) {
-        if (this.ctrl != null) try {
-            ctrl.setEnabled(enabled);
-        } catch (BBjException e) {
-            e.printStackTrace();
-        }
-        return this;
-    }
+    // @Override
+    // public TextComboBox setEnabled(boolean enabled) {
+    //     if (this.ctrl != null) try {
+    //         ctrl.setEnabled(enabled);
+    //     } catch (BBjException e) {
+    //         e.printStackTrace();
+    //     }
+    //     return this;
+    // }
 
-    @Override
-    public boolean isEnabled() {
-        if (this.ctrl != null) try {
-            return ctrl.isEnabled();
-        } catch (BBjException e) {
-            e.printStackTrace();
-        }
-        return true;
-    }
+    // @Override
+    // public boolean isEnabled() {
+    //     if (this.ctrl != null) try {
+    //         return ctrl.isEnabled();
+    //     } catch (BBjException e) {
+    //         e.printStackTrace();
+    //     }
+    //     return true;
+    // }
 }
